+++ service/service-share/src/main/java/com/demo/service/share/DemoService.java
@@ -6,7 +6,7 @@ public interface DemoService {
 
     int deleteByPrimaryKey(Long id);
 
-    int insert(DemoBO demoBO);
+    Long insert(DemoBO demoBO);
 
     int insertSelective(DemoBO demoBO);
